Deprecate Network, Authentication, Scripting Data
