+++ Source/Kesmai.WorldForge/Core.xaml
@@ -5,7 +5,7 @@
 
              ShutdownMode="OnMainWindowClose"
 			 
-             StartupUri="UI\Windows\AuthenticationWindow.xaml">
+             StartupUri="UI\ApplicationWindow.xaml">
     <Application.Resources>
         <presenters:ApplicationPresenter x:Key="applicationPresenter" />
     </Application.Resources>
