+++ Source/Kesmai.WorldForge/Kesmai.WorldForge.csproj
@@ -93,9 +93,6 @@
       <Reference Include="DotNext">
         <HintPath>..\..\Dependencies\DotNext.dll</HintPath>
       </Reference>
-      <Reference Include="Lidgren.Network, Version=2012.1.7.0, Culture=neutral, PublicKeyToken=null">
-        <HintPath>..\..\Dependencies\Lidgren.Network.dll</HintPath>
-      </Reference>
     </ItemGroup>
 
     <ItemGroup>
@@ -141,7 +138,6 @@
             <GarbageFile Include="DotNetZip.dll" />
             <GarbageFile Include="Humanizer.dll" />
             <GarbageFile Include="ICSharpCode.AvalonEdit.dll" />
-            <GarbageFile Include="Lidgren.Network.dll" />
             <GarbageFile Include="Microsoft.Bcl.AsyncInterfaces.dll" />
             <GarbageFile Include="Microsoft.CodeAnalysis.AnalyzerUtilities.dll" />
             <GarbageFile Include="Microsoft.CodeAnalysis.CSharp.dll" />
