+++ Source/Kesmai.WorldForge/UI/ApplicationWindow.xaml
@@ -32,8 +32,6 @@
                 <MenuItem Header="_Open Segment" Command="{Binding OpenSegmentCommand}" />
                 <MenuItem Header="C_lose Segment"  Command="{Binding CloseSegmentCommand}" />
                 <Separator />
-                <MenuItem Header="Compile Segment"  Command="{Binding CompileSegmentCommand}" />
-                <Separator />
                 <MenuItem Header="_Save Segment"  Command="{Binding SaveSegmentCommand}">
                     <MenuItem.Icon>
                         <Image Source="../Resources/Save.png" Width="16" Height="16" />
