--- Source/Kesmai.WorldForge/UI/Windows/AuthenticationWindow.xaml
@@ -1,54 +1,0 @@
-<Window x:Class="Kesmai.WorldForge.UI.AuthenticationWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        
-        Title = "WorldForge - Login"
-        
-        mc:Ignorable="d"
-        
-        SizeToContent="WidthAndHeight"
-        ResizeMode="NoResize"
-        WindowStyle="None"
-        
-        WindowStartupLocation="CenterScreen">
-    <Grid Margin="5" Width="310">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*"></RowDefinition>
-            <RowDefinition Height="50"></RowDefinition>
-        </Grid.RowDefinitions>
-        
-        <Image Grid.Row="0" 
-               Source="pack://application:,,,/Kesmai.WorldForge;component/Resources/WorldForge-Logo.png"/>
-        
-        <ProgressBar Minimum="0" Maximum="100" x:Name="_progress"
-                     Grid.Row="1">
-        </ProgressBar>
-        
-        <TextBlock Text="Loading.." VerticalAlignment="Center" Margin="5,0,5,0" 
-                   FontFamily="Tahoma" FontWeight="Bold"
-                   x:Name="_status" Grid.Row="1" />
-        
-        <StackPanel Orientation="Horizontal" Margin="2" HorizontalAlignment="Right"
-                    Grid.Row="1">
-            <StackPanel.Triggers>
-                <EventTrigger RoutedEvent="Rectangle.Loaded">
-                    <BeginStoryboard>
-                        <Storyboard Name="waitingAnimation" RepeatBehavior="Forever">
-                            <DoubleAnimation Storyboard.TargetName="Dot1" BeginTime="0:0:0.1" Storyboard.TargetProperty="Opacity" From="0" To="1" AutoReverse="True" Duration="0:0:0.3"/>
-                            <DoubleAnimation Storyboard.TargetName="Dot2" BeginTime="0:0:0.2" Storyboard.TargetProperty="Opacity" From="0" To="1" AutoReverse="True" Duration="0:0:0.3"/>
-                            <DoubleAnimation Storyboard.TargetName="Dot3" BeginTime="0:0:0.3" Storyboard.TargetProperty="Opacity" From="0" To="1" AutoReverse="True" Duration="0:0:0.3"/>
-                        </Storyboard>
-                    </BeginStoryboard>
-                </EventTrigger>
-            </StackPanel.Triggers>
-            
-            <Ellipse Name="Dot1" Fill="Yellow" Opacity="0" Width="8" Height="8"/>
-            <Ellipse Name="Dot2" Fill="Magenta" Opacity="0" Width="8" Height="8" Margin="2,0,0,0"/>
-            <Ellipse Name="Dot3" Fill="Cyan" Opacity="0" Width="8" Height="8" Margin="2,0,0,0"/>
-        </StackPanel>
-    </Grid>
-    
-   
-</Window>
