--- Source/Kesmai.WorldForge/UI/Windows/CompileWindow.xaml
@@ -1,50 +1,0 @@
-﻿<Window x:Class="Kesmai.WorldForge.UI.Windows.CompileWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        
-        Title = "Compile"
-        
-        mc:Ignorable="d"
-        
-        SizeToContent="WidthAndHeight"
-        ResizeMode="NoResize"
-        WindowStyle="ToolWindow"
-        
-        WindowStartupLocation="CenterOwner">
-    <Grid Margin="5" Width="710">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*"></RowDefinition>
-            <RowDefinition Height="*"></RowDefinition>
-        </Grid.RowDefinitions>
-        
-        <TextBlock Text=".. Compiling .." VerticalAlignment="Center" HorizontalAlignment="Center" Margin="5" 
-                   FontFamily="Tahoma" FontWeight="Bold"
-                   x:Name="_status" Grid.Row="0" />
-        
-        <DataGrid Height="300" Width="700" Grid.Row="1" Margin="5"
-                  HorizontalAlignment="Left"  
-                  VerticalAlignment="Top"
-                  FontFamily="Tahoma" FontSize="10" IsReadOnly="True"
-                  x:Name="_dataGrid"  RowHeight="30" ColumnWidth="100" AutoGenerateColumns="false">
-            <DataGrid.Columns>
-                <DataGridTextColumn Width="90" Binding="{Binding Class}" Header="Class"></DataGridTextColumn>
-                <DataGridTextColumn Width="300" Binding="{Binding Error}" Header="Error">
-                    <DataGridTextColumn.ElementStyle>
-                        <Style>                            
-                            <Setter Property="TextBlock.TextWrapping" Value="Wrap" />
-                        </Style>
-                    </DataGridTextColumn.ElementStyle>
-                </DataGridTextColumn>
-                <DataGridTextColumn Width="300" Binding="{Binding Syntax}" Header="Syntax">
-                    <DataGridTextColumn.ElementStyle>
-                        <Style>                            
-                            <Setter Property="TextBlock.TextWrapping" Value="Wrap" />
-                        </Style>
-                    </DataGridTextColumn.ElementStyle>
-                </DataGridTextColumn>
-            </DataGrid.Columns>
-        </DataGrid>
-    </Grid>
-</Window>
